Frontend fixes
