--- .env.local
@@ -1,3 +1,0 @@
-YOUTUBE_API_KEY=AIzaSyCGuF9vosG65GuVpdlJxmxEpgCR1BgYdFw
-
-
