--- README.md
@@ -1,4 +1,0 @@
-# playlistify
-
-When you want to learn something, you go on youtube. But searching there is a daunting task because it feels like there are so many videos as you scroll through them all, which one should you watch? And even if you know one or two that are really good, how can you find them automatically?  	                       
-         	So here’s an app that might help you find the most relevant playlist for your needs. The purpose of this website is to help you find the most relevant videos for your interests and you can even organise your own personalised playlist here
