+++ backend/.env.local
@@ -1,0 +1,3 @@
+YOUTUBE_API_KEY=AIzaSyCGuF9vosG65GuVpdlJxmxEpgCR1BgYdFw
+
+
