--- package.json
@@ -1,9 +1,0 @@
-{
-  "dependencies": {
-    "cors": "^2.8.5",
-    "dotenv": "^16.0.3",
-    "express": "^4.18.2",
-    "googleapis": "^112.0.0",
-    "react-axios": "^2.0.6"
-  }
-}
