--- src/Categories.js
@@ -1,17 +1,0 @@
-import React from 'react'
-import Categorynav from './Components/Categorynav'
-import Categorymain from './Components/Categorymain'
-
-function Categories() {
-  return (
-
-    <div>
-        
-  <Categorynav/>
-  <Categorymain/>
-  
-    </div>
-  )
-}
-
-export default Categories
