--- src/Components/Categories.css
@@ -1,56 +1,0 @@
-
-.dance-categories1{
-    display: flex;
-    justify-content: space-evenly;
-    color: black;
-    margin: 100px 0px 0px 0px;
-    
-}
-#text{
-    text-decoration: none;
-}
-#block{
-    display: block;
-    /* flex-direction: row; */
-}
-.blurry-background {
-    backdrop-filter: blur(10px);
-    padding: 20px;
-    /* background-image: url(dance2); */
-    height:250px;
-    width: 200px;
-    background-size: cover;
-    
-  }
-  
-  h2 {
-    color: #fff;
-  }
-  .kathak-dance{
-    height: 100pxx;
-    width: 500px;
-    /* backdrop-filter: blur(10px); */
-
-  }
-  
-  .blurry-background2{
-    height: 250px;
-    width: 235px;
-  }
-
-  .dance-categories2{
-    display: flex;
-    justify-content: space-evenly;
-    color: black;
-    margin-top: 35px;
-  }
-
-  .kathak-api{
-    margin: 34px 40px 14px 58px;
-
-  }
-  
-  .kathak-api-image{
-    height:300px;
-    
-  }
