--- src/Components/Homepage.css
@@ -1,40 +1,0 @@
-body{
-  margin: 0px;
-  padding: 0px;
-  background-color: #F2D8FF;
-}
-
-
-.homepage{
-  display: inline;
-  flex-direction: row;
-}
-  
-
-.musictea{
-  /* display: flex; */
-  flex-direction: row;
-  height: 250px;
-  width: auto;
-  margin: 40px 0px 200px 15px;
-
-}
-
-.guy{
-  /* display: flex; */
-  flex-direction: row;
-  height: 300px;
-  width: 500px;
-  margin: 0px 0px 280px -25px;
-  position: relative;
-  z-index: 2;
-  
-}
-
-.dance{
-  /* display: flex; */
-  flex-direction: row;
-  height: 630px;
-  width: 649px;;
-  margin: 0px 0px 210px -20px;
-}
