--- src/Components/Navbar.css
@@ -1,76 +1,0 @@
-.logo{
-    width: 20vw;
-    
-}
-
-nav{
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    background-color: #F2D8FF
-    ;
-}
-
-#flex{
-    display: flex;
-    flex-direction: row;
-    justify-content: space-around;
-    width:  30vw;
-
-}
-
-.carousel-image{
-    width :100%;
-    height: 600px;
-}
-
-.about-image {
-    margin: 0px 262px 0px 420px;
-    height: 450px;
-    width: 500px;
-}
-
-.content-div{
-    padding: 112px 0px 0px 0px;
-    align-items: center;
-}
-
-.content-text{
-   margin: 0px 0px 0px 213px;
-}
-
-.card {
-    width: 300px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    padding: 20px;
-    margin: 20px;
-  }
-  
-  .card img {
-    width: 100%;
-    height: auto;
-    margin-bottom: 10px;
-  }
-
-  
-.cards-container{
-    padding: 112px 0px 0px 0px;
-    display: flex;
-    justify-content: space-evenly;
-}
-
-.About-heading{
-    margin: 75px 0px 0px 650px;
-}
-
-#para{
-    width: 300px;
-
-}
-
-#dance-button{
-    background-color: black;
-    color: #ccc;
-  margin: 20px 0px 0px 110px;
-}
