--- src/Kathak.js
@@ -1,15 +1,0 @@
-import React from 'react'
-import Kathaknav from './Components/Kathaknav'
-import Kathakmain from './Components/Kathakmain'
-
-function Kathak() {
-  return (
-    <div>
-        <Kathaknav/>
-        <Kathakmain />
-
-    </div>
-  )
-}
-
-export default Kathak
