--- src/index.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-
-
-
-import App from './App';
-
-
-
-
-const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-
-    <App />
-);
-
